Throttle "Ignoring late packet" messages
Previously, each late packet was logged because late packets was assumed
to be a rare occurrence.  Unfortunately, due to not-yet-understood
circumstances, the HERA X engines sometimes receive "late" packets from
some of the SNAPs, sometimes many many late packets from some of the
SNAPs.  Logging all of these late packets causes the log files to grow
until eventually the file system is filled up, which further impacts
observing.

These messages are now logged in a still generous, but much more
constrained quantity.  Up to 5,000 of these "late packet" messages will
be logged within an hour from the first such log message.  After that
hour has elapsed, the next such log message will start a new hour long
window.  This will result in a max of 120,000 such log messages in a 24
hour period (per X engine), which is far fewer than the 3+ million such
log messages per X engine that led to a full file system.

The threshold of 5,000 was chosen because there seemed to be some
periodicity in the neighborhood of 1024 occurrences.  The limit of 5,000
will allow for the capture of several such cycles per "burst".